Move github client to its own entitiy [WIP]
